This PR contains a major improvement to the FeatureLinkerUnlabeledQT which I have found to perform rather slow and consume huge amounts of memory when I tried to run it. These problems are addressed in this PR.

The main problem with FeatureLinkerUnlabeledQT was that the algorithm basically kept around a lot of data structures that grew quadratically with the number of input features. Specifically, `distances_` was basically a hash of all distances between all features that were ever needed in the past which is `O(n^2 * k^2)` where *n* is the number of features per map and *k* is the number of maps (thus total number of features `m = n* k`). Similarly, `QTCluster` was a data structure created for each feature in each map (a total of `m` instances) and each instance held a list of *all* its neighbors in *all* the other maps, so this would grow with `O(c * m^2) = O( n^2 * k^2)` where `c` is a constant. Note that of course not every cluster node has a list of every other cluster node but only the immediate neighbours, so the constant factor `c` is quite small but the growth is of course still `O( n^2 * k^2)`.

I have modified these two data structures, eliminated `distances_` completely and modified `QTCluster` such that it only stores the *best* feature from every map, thus reducing memory complexity to `O(n * k^2)` which showed big improvements on my machine. Unfortunately, memory is still going to grow quadratically with the number of input maps. Even doing this change turned out to be non-trivial as there is one corner-case when using identifications where the correct cluster depends on *all* neighbors. For this case, I now temporary store all neighbors and delete the map again after computing the correct cluster.

Additionally I have introduced an optional partitioning by *m/z* as often the distance between features in m/z domain can be quite large and way larger than the maximal distance to be considered for a cluster. This means that the problem can be partitioned and each partition can be solved independently.

Using the partitioning, even runtime goes down. With one partition, runtime is constant or somewhat larger. This is despite the fact that I remove some speed-optimization and some work now needs to be done multiple times (e.g. now every time a neighbor gets removed from a cluster, all potential neighbors need to be re-investigated and everything needs to be recomputed).

Some data on a dataset of 7 feature maps with  about 10k features each that were re-used multiple times:

Dataset of 7 maps:
- before: 616 MB memory usage and 112 seconds runtime
- after (1 partition): 88 MB of memory and 64 seconds runtime
- after (30 partitions): 43 MB of memory and 9.5 seconds runtime

so this is an 10x improvement in runtime and a 14x improvement in memory usage

Dataset of 28 maps (re-using the same maps 4 times)
- before: 2.4 GB / 5:26 min
- after (45 partitions): 127 MB memory usage / 65 seconds runtime


Now memory usage seems to scale (almost) linearly with the number of maps and the good thing is that the algorithm now seems to scale to dozens of maps, if not hundreds (it starts to get slow at some point). This is all computed using 30 partitions:

- 7 maps: 43 MB of memory
- 28 maps: 127 MB of memory
- 56 maps: 276.5 MB of memory
- 112 maps: 666.5 MB of memory


caveats: this is now only tested on raw maps without any id data (currently I do not have any such data) and -- surprise surprise -- there has not been a test with the `use_identifications` enabled. I have tried to manage this case also carefully but I would be very glad if somebody could test this on a larger dataset and confirm that  the results are the same. I have also created a test data where `use_identifications` produces different results and this is now tested in a toy example.