@hendrikweisser 2015-09-22
Hannes, thanks a lot for this. I'll do a code review as soon as I get more time for it.

Some thoughts based on your comments above:

> The main problem with FeatureLinkerUnlabeledQT was that the algorithm basically kept around a lot of data structures that grew quadratically with the number of input features.

The original code tried to optimize speed at the cost of memory. I still think that it's beneficial to sacrifice memory to save runtime here, as long as the memory consumption doesn't preclude users from analysing large datasets on relatively powerful hardware. So, I wonder what the effect would be of keeping the hashing of computed distances (`distances_`) to save you from having to recompute them. Would it be possible to decide earlier on which distances you might have to recompute and store only those? Alternatively, if the effect on runtime is considerable, perhaps a flag could be introduced to leave the memory vs. runtime decision to the user.

> Even doing this change turned out to be non-trivial as there is one corner-case when using identifications where the correct cluster depends on all neighbors. For this case, I now temporary store all neighbors and delete the map again after computing the correct cluster.

Does "delete the map again" mean you potentially have to recompute things?

> Additionally I have introduced an optional partitioning by m/z as often the distance between features in m/z domain can be quite large and way larger than the maximal distance to be considered for a cluster. This means that the problem can be partitioned and each partition can be solved independently.

This sounds very good, but shouldn't the HashGrid already provide partitioning by RT and m/z? Am I wrong, or is that for some reason not sufficient? If so, can you automatically set the partition size based on the maximum allowed m/z difference?

> there has not been a test with the `use_identifications` enabled

It would of course be good to have a tool test covering this, but note that all the different cases should be checked in the QTClusterFinder class test (https://github.com/OpenMS/OpenMS/blob/develop/src/tests/class_tests/openms/source/QTClusterFinder_test.cpp#L178 and below). If those tests still pass, everything should be fine.