@hroest 2015-09-22
> The original code tried to optimize speed at the cost of memory. I still think that it's beneficial to sacrifice memory to save runtime here, as long as the memory consumption doesn't preclude users from analysing large datasets on relatively powerful hardware. So, I wonder what the effect would be of keeping the hashing of computed distances (`distances_`) to save you from having to recompute them. Would it be possible to decide earlier on which distances you might have to recompute and store only those? Alternatively, if the effect on runtime is considerable, perhaps a flag could be introduced to leave the memory vs. runtime decision to the user.

Yes, I can totally see that but in my case it turned out that about 40-50% of the memory was used for the `distances_` map, storing old distances and I did not find any decrease in runtime once I disabled caching of the distances. So it does not seem to be the bottleneck for computation and removing it also makes the code run faster. So maybe storing all of these in a map and looking them up is slower than recomputing? I only tested this with the default settings, if somebody wanted to use a different exponent or has a really expensive distance measure, then it may be different of course.

> Does "delete the map again" mean you potentially have to recompute things?

Yes, unfortuntately I have to. Every time we create a new consensus feature, we have to iterate through all clusters that were potentially affected by this operation and remove the elements from this cluster. We already did this before, but now the operation is, at least potentially in some cases, more expensive. If the removed elements were among the "best elements" for the affected cluster, we basically have to throw out the whole computation, find all potential neighbors for the affected cluster *again* and add them to the cluster again (of course omitting the already removed ones). Have a look at `addClusterElements_` for this. This sounds *terribly* expensive -- surprisingly it turns out to be not that expensive after all. It may be that in the end not so many clusters are actually affected since only in few cases would we remove somebody elses "best elements". Or maybe keeping fewer elements per cluster and in few cases do a lot of work balances the previous solution of keeping many elements per cluster and in all cases do a little bit of work? (the "little" bit of work being that we almost always needed to perform some update when removing elements whereas now we rarely need to do that). At least in my test data, the patch improves runtime so it does not seem to balance out.
This is basically done here: https://github.com/OpenMS/OpenMS/pull/1610/files#diff-067fc2140904180fda6b08dba8c57cc9R366

> This sounds very good, but shouldn't the HashGrid already provide partitioning by RT and m/z? Am I wrong, or is that for some reason not sufficient? If so, can you automatically set the partition size based on the maximum allowed m/z difference?

yes, it does, but this did not save you any memory at the time. Now we actually compute clusters only for all features between 200-240 m/z, then for 240-260 m/z and so on. This means we have *p* times fewer cluster in our list of clusters (where *p* = number of partitions) and since the `QTCluster` objects take most of the memory, peak memory usage goes down. Also it seems runtime seems to go up since some operations (like iterating over all clusters) take `O( n*k / p)` time instead of `O(n*k)`. It may also be that keeping things more local is faster for other reasons (cache locality etc), I dont really know, this is just speculation. However, in some cases (for the dataset of 7-28 files) I see 3x to 6x speed improvement when turning on the partitioning.

> It would of course be good to have a tool test covering this, but note that all the different cases should be checked in the QTClusterFinder class test (https://github.com/OpenMS/OpenMS/blob/develop/src/tests/class_tests/openms/source/QTClusterFinder_test.cpp#L178 and below). If those tests still pass, everything should be fine.

ok, great! I worried a bit that I may have broken the `use_identifications`  option but it seems that the `QTClusterFinder_test.cpp` still works. I have now also added a TOPP test using the identifications.