@timosachsenberg 2015-09-24
Sounds great. I like the partitioning - the next extension would be to do a DBScan clustering first and then use the resulting clusters as an initial partitioning (and also to resolve all clusters within the QT) ;)